backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2
Werkzeug==3.0.1 